requirements.txt
annotated-types==0.7.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
typing-inspection==0.4.0
typing_extensions==4.13.0
